feat: basic lint and test workflow
Leverages the install-php base workflow and is itself a consumable
workflow. The actual steps for linting and testing need to be
added, based on how that will be done: will it be defined by a
composer script, allowing for flexibility; or will it be assumed
based on a conventional directory structure?